requirements.txt
chainlit
langchain
langchain-google-genai
langchain-community>
python-dotenv
faiss-cpu
pypdf
google-generativeai
fastapi
uvicorn
pydantic
python-socketio
aiohttp
python-docx
unstructured
sentence-transformers
torch